requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
boto3==1.26.129
botocore==1.29.129
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
cssselect2==0.7.0
Django==4.1.1
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-rest-framework==0.1.0
django-storages==1.13.2
django-tinymce==3.6.1
djangorestframework==3.13.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
lxml==4.9.1
mysqlclient==2.1.1
oscrypto==1.3.0
Pillow==9.2.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
webencodings==0.5.1
